Move publishersstyle to global scope
