+++ src/components/Epub/Settings/index.ts
@@ -3,4 +3,3 @@
 export * from "./StatefulColumns";
 export * from "./StatefulLayout";
 export * from "./StatefulTheme";
-export * from "./StatefulPublisherStyles";
